ADDED SEARCH FUNCTION FOR USERS AND MODIFIED FOOTER
I noticed in the blog section if a user likes a post it only increments
the count for the first blog post so I added a data attribute storing the
id of the blog and using the aria label for the like button query the
like count so that it shows the increments properly.

I also added a next button to the view-blog section so that the user can
have the option to view the next blog, the contents to be loaded is
randomized with the condition that it doesn't match with the current
blog id

I also changed the carousel to display the top liked posts limited to
only 3 posts

In the footer I included a message section in where the user can submit
a concern that will be stored in the database, and the admin can view the
concern. This function is still in progress